Update acc test
